+++ A32NX/html_ui/Pages/VCockpit/Instruments/Airliners/A320_Neo/CDU/A320_Neo_CDU_MainDisplay.js
@@ -545,7 +545,7 @@ class A320_Neo_CDU_MainDisplay extends FMCMainDisplay {
             if (this.messageQueue[0][2](this)) {
                 this.messageQueue.splice(0, 1);
                 this._inOutElement.className = "white";
-                this.inOut = this.lastUserInput;
+                this.lastUserInputToScratchpad();
                 return this.tryShowMessage();
             }
             if (!this.isDisplayingErrorMessage) {
@@ -570,7 +570,7 @@ class A320_Neo_CDU_MainDisplay extends FMCMainDisplay {
                 this.messageQueue.splice(i, 1);
                 if (i === 0 && this.isDisplayingTypeTwoMessage) {
                     this._inOutElement.className = "white";
-                    this.inOut = this.lastUserInput;
+                    this.lastUserInputToScratchpad();
                 }
                 break;
             }